Fix theme cycle
