+++ frontend/app/package.json
@@ -212,8 +212,10 @@
     "axios-mock-adapter": "^1.21.2",
     "babel-cli": "^6.26.0",
     "babel-plugin-emotion": "^11.0.0",
+    "babel-plugin-module-resolver": "^5.0.0",
     "babel-plugin-tsconfig-paths-module-resolver": "^1.0.4",
     "buffer": "^6.0.3",
+    "circular-dependency-plugin": "^5.2.2",
     "cypress": "^9.7.0",
     "cypress-file-upload": "^5.0.8",
     "cypress-image-snapshot": "^4.0.1",
@@ -250,6 +252,7 @@
     "typescript": "^4.9.5",
     "util": "^0.12.5",
     "webpack": "5.75.0",
+    "webpack-bundle-analyzer": "^4.8.0",
     "webpack-cli": "^5.0.1"
   },
   "resolutions": {
