+++ frontend/lib/package.json
@@ -6,6 +6,7 @@
   "homepage": "./",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
+  "sideEffects": false,
   "scripts": {
     "build": "yarn build:src && yarn build:autogen && yarn build:asset && yarn build:lib-type",
     "test":  "env NODE_OPTIONS=--max_old_space_size=8192 jest --env=../sharedConfigs/jest/jsdom-polyfill-env.js --runTestsByPath ${TESTPATH}",
