+++ frontend/lib/src/theme/baseTheme/index.ts
@@ -28,7 +28,7 @@ import {
   spacing,
   zIndices,
 } from "src/theme/primitives"
-import { createEmotionColors } from "src/theme/utils"
+import { createEmotionColors } from "src/theme/getColors"
 import genericColors from "./themeColors"
 
 const baseTheme = {
