+++ frontend/lib/src/theme/themeConfigs.ts
@@ -25,7 +25,6 @@ import base from "./baseTheme"
 import light from "./lightTheme"
 import dark from "./darkTheme"
 import { ThemeConfig } from "./types"
-import { AUTO_THEME_NAME, getSystemTheme } from "./utils"
 
 export const baseTheme: ThemeConfig = {
   name: "base",
@@ -50,15 +49,3 @@ export const lightTheme: ThemeConfig = {
   basewebTheme: lightBaseUITheme,
   primitives: lightThemePrimitives,
 }
-
-export const createAutoTheme = (): ThemeConfig => ({
-  ...getSystemTheme(),
-  name: AUTO_THEME_NAME,
-})
-
-// Update auto theme in case it has changed
-export const createPresetThemes = (): ThemeConfig[] => [
-  createAutoTheme(),
-  lightTheme,
-  darkTheme,
-]
